\name{proteinProfiles-package}

\docType{package}

\alias{proteinProfiles}
\alias{proteinProfiles-package}

\title{proteinProfiles package}

\description{
  Significance assessment for distance measures of time-course protein profiles
}

\details{
  The package is published under the GPL-3 license.
}

\author{
  Julian Gehring

  Maintainer: Julian Gehring <julian.gehring@embl.de>
}

\seealso{
  filterFeatures, grepAnnotation, profileDistance, plotProfileDistance
}

\keyword{package}
\keyword{htest}
\keyword{distribution}